+++ translations/location_history_visualizer_de.ts
@@ -20,6 +20,10 @@
         <translation>&amp;Einstellungen</translation>
     </message>
     <message>
+        <source>Map display...</source>
+        <translation>Kartendarstellung...</translation>
+    </message>
+    <message>
         <source>&amp;Help</source>
         <translation>&amp;Hilfe</translation>
     </message>
@@ -34,6 +38,18 @@
     <message>
         <source>No file loaded</source>
         <translation>Keine Datei geladen</translation>
+    </message>
+    <message>
+        <source>Counts</source>
+        <translation>Anzahl</translation>
+    </message>
+    <message>
+        <source>In file</source>
+        <translation>In der Datei</translation>
+    </message>
+    <message>
+        <source>Visible</source>
+        <translation>Sichtbar</translation>
     </message>
     <message>
         <source>Date</source>
@@ -196,6 +212,10 @@
         <translation>%1 Punkte angezeigt</translation>
     </message>
     <message>
+        <source>%1 of %2 points shown</source>
+        <translation>%1 von %2 Punkten angezeigt</translation>
+    </message>
+    <message>
         <source>Open Timeline JSON</source>
         <translation>Timeline JSON öffnen</translation>
     </message>
@@ -241,6 +261,25 @@
     <message>
         <source>&lt;b&gt;GitHub repository:&lt;/b&gt; %1</source>
         <translation>&lt;b&gt;GitHub-Repository:&lt;/b&gt; %1</translation>
+    </message>
+</context>
+<context>
+    <name>LocationHistory::MapDisplayDialog</name>
+    <message>
+        <source>Map display</source>
+        <translation>Kartendarstellung</translation>
+    </message>
+    <message>
+        <source> px</source>
+        <translation> px</translation>
+    </message>
+    <message>
+        <source>Point size</source>
+        <translation>Punktgröße</translation>
+    </message>
+    <message>
+        <source>Maximum drawn points</source>
+        <translation>Maximal gezeichnete Punkte</translation>
     </message>
 </context>
 </TS>