+++ translations/location_history_visualizer_es.ts
@@ -20,6 +20,10 @@
         <translation>&amp;Configuración</translation>
     </message>
     <message>
+        <source>Map display...</source>
+        <translation>Visualización del mapa...</translation>
+    </message>
+    <message>
         <source>&amp;Help</source>
         <translation>A&amp;yuda</translation>
     </message>
@@ -34,6 +38,18 @@
     <message>
         <source>No file loaded</source>
         <translation>Ningún archivo cargado</translation>
+    </message>
+    <message>
+        <source>Counts</source>
+        <translation>Recuento</translation>
+    </message>
+    <message>
+        <source>In file</source>
+        <translation>En el archivo</translation>
+    </message>
+    <message>
+        <source>Visible</source>
+        <translation>Visibles</translation>
     </message>
     <message>
         <source>Date</source>
@@ -196,6 +212,10 @@
         <translation>%1 puntos mostrados</translation>
     </message>
     <message>
+        <source>%1 of %2 points shown</source>
+        <translation>%1 de %2 puntos mostrados</translation>
+    </message>
+    <message>
         <source>Open Timeline JSON</source>
         <translation>Abrir JSON de Timeline</translation>
     </message>
@@ -241,6 +261,25 @@
     <message>
         <source>&lt;b&gt;GitHub repository:&lt;/b&gt; %1</source>
         <translation>&lt;b&gt;Repositorio de GitHub:&lt;/b&gt; %1</translation>
+    </message>
+</context>
+<context>
+    <name>LocationHistory::MapDisplayDialog</name>
+    <message>
+        <source>Map display</source>
+        <translation>Visualización del mapa</translation>
+    </message>
+    <message>
+        <source> px</source>
+        <translation> px</translation>
+    </message>
+    <message>
+        <source>Point size</source>
+        <translation>Tamaño de punto</translation>
+    </message>
+    <message>
+        <source>Maximum drawn points</source>
+        <translation>Máximo de puntos dibujados</translation>
     </message>
 </context>
 </TS>