+++ translations/location_history_visualizer_pt.ts
@@ -20,6 +20,10 @@
         <translation>&amp;Definições</translation>
     </message>
     <message>
+        <source>Map display...</source>
+        <translation>Visualização do mapa...</translation>
+    </message>
+    <message>
         <source>&amp;Help</source>
         <translation>&amp;Ajuda</translation>
     </message>
@@ -34,6 +38,18 @@
     <message>
         <source>No file loaded</source>
         <translation>Nenhum ficheiro carregado</translation>
+    </message>
+    <message>
+        <source>Counts</source>
+        <translation>Contagens</translation>
+    </message>
+    <message>
+        <source>In file</source>
+        <translation>No ficheiro</translation>
+    </message>
+    <message>
+        <source>Visible</source>
+        <translation>Visíveis</translation>
     </message>
     <message>
         <source>Date</source>
@@ -196,6 +212,10 @@
         <translation>%1 pontos mostrados</translation>
     </message>
     <message>
+        <source>%1 of %2 points shown</source>
+        <translation>%1 de %2 pontos mostrados</translation>
+    </message>
+    <message>
         <source>Open Timeline JSON</source>
         <translation>Abrir JSON Timeline</translation>
     </message>
@@ -241,6 +261,25 @@
     <message>
         <source>&lt;b&gt;GitHub repository:&lt;/b&gt; %1</source>
         <translation>&lt;b&gt;Repositório GitHub:&lt;/b&gt; %1</translation>
+    </message>
+</context>
+<context>
+    <name>LocationHistory::MapDisplayDialog</name>
+    <message>
+        <source>Map display</source>
+        <translation>Visualização do mapa</translation>
+    </message>
+    <message>
+        <source> px</source>
+        <translation> px</translation>
+    </message>
+    <message>
+        <source>Point size</source>
+        <translation>Tamanho dos pontos</translation>
+    </message>
+    <message>
+        <source>Maximum drawn points</source>
+        <translation>Máximo de pontos desenhados</translation>
     </message>
 </context>
 </TS>