+++ translations/location_history_visualizer_ru.ts
@@ -20,6 +20,10 @@
         <translation>&amp;Настройки</translation>
     </message>
     <message>
+        <source>Map display...</source>
+        <translation>Отображение карты...</translation>
+    </message>
+    <message>
         <source>&amp;Help</source>
         <translation>&amp;Справка</translation>
     </message>
@@ -34,6 +38,18 @@
     <message>
         <source>No file loaded</source>
         <translation>Файл не загружен</translation>
+    </message>
+    <message>
+        <source>Counts</source>
+        <translation>Количество</translation>
+    </message>
+    <message>
+        <source>In file</source>
+        <translation>В файле</translation>
+    </message>
+    <message>
+        <source>Visible</source>
+        <translation>Видимые</translation>
     </message>
     <message>
         <source>Date</source>
@@ -196,6 +212,10 @@
         <translation>Показано точек: %1</translation>
     </message>
     <message>
+        <source>%1 of %2 points shown</source>
+        <translation>Показано %1 из %2 точек</translation>
+    </message>
+    <message>
         <source>Open Timeline JSON</source>
         <translation>Открыть JSON Timeline</translation>
     </message>
@@ -241,6 +261,25 @@
     <message>
         <source>&lt;b&gt;GitHub repository:&lt;/b&gt; %1</source>
         <translation>&lt;b&gt;Репозиторий GitHub:&lt;/b&gt; %1</translation>
+    </message>
+</context>
+<context>
+    <name>LocationHistory::MapDisplayDialog</name>
+    <message>
+        <source>Map display</source>
+        <translation>Отображение карты</translation>
+    </message>
+    <message>
+        <source> px</source>
+        <translation> px</translation>
+    </message>
+    <message>
+        <source>Point size</source>
+        <translation>Размер точек</translation>
+    </message>
+    <message>
+        <source>Maximum drawn points</source>
+        <translation>Максимум отображаемых точек</translation>
     </message>
 </context>
 </TS>