+++ translations/location_history_visualizer_tr.ts
@@ -20,6 +20,10 @@
         <translation>&amp;Ayarlar</translation>
     </message>
     <message>
+        <source>Map display...</source>
+        <translation>Harita görünümü...</translation>
+    </message>
+    <message>
         <source>&amp;Help</source>
         <translation>&amp;Yardım</translation>
     </message>
@@ -34,6 +38,18 @@
     <message>
         <source>No file loaded</source>
         <translation>Dosya yüklenmedi</translation>
+    </message>
+    <message>
+        <source>Counts</source>
+        <translation>Sayılar</translation>
+    </message>
+    <message>
+        <source>In file</source>
+        <translation>Dosyada</translation>
+    </message>
+    <message>
+        <source>Visible</source>
+        <translation>Görünür</translation>
     </message>
     <message>
         <source>Date</source>
@@ -196,6 +212,10 @@
         <translation>%1 nokta gösteriliyor</translation>
     </message>
     <message>
+        <source>%1 of %2 points shown</source>
+        <translation>%2 noktadan %1 tanesi gösteriliyor</translation>
+    </message>
+    <message>
         <source>Open Timeline JSON</source>
         <translation>Timeline JSON aç</translation>
     </message>
@@ -241,6 +261,25 @@
     <message>
         <source>&lt;b&gt;GitHub repository:&lt;/b&gt; %1</source>
         <translation>&lt;b&gt;GitHub deposu:&lt;/b&gt; %1</translation>
+    </message>
+</context>
+<context>
+    <name>LocationHistory::MapDisplayDialog</name>
+    <message>
+        <source>Map display</source>
+        <translation>Harita görünümü</translation>
+    </message>
+    <message>
+        <source> px</source>
+        <translation> px</translation>
+    </message>
+    <message>
+        <source>Point size</source>
+        <translation>Nokta boyutu</translation>
+    </message>
+    <message>
+        <source>Maximum drawn points</source>
+        <translation>Çizilecek en fazla nokta</translation>
     </message>
 </context>
 </TS>